fix some warnings where preprocessor expected a newline
