+++ MeasureResults/MeasureResults.hpp
@@ -1,0 +1,5 @@
+void setUpResults(std::map<std::string, double>&);
+
+void messureResults(size_t start, const size_t, std::map<std::string, double>&, const std::map<std::string, Results*>&);
+
+void displayResults(std::map<std::string, double>&);
