+++ MeasureResults/MeasureResults.vcxproj.filters
@@ -21,6 +21,9 @@
     <ClInclude Include="Results.hpp">
       <Filter>Header Files</Filter>
     </ClInclude>
+    <ClInclude Include="MeasureResults.hpp">
+      <Filter>Header Files</Filter>
+    </ClInclude>
   </ItemGroup>
   <ItemGroup>
     <ClCompile Include="main.cpp">
@@ -30,6 +33,9 @@
       <Filter>Source Files</Filter>
     </ClCompile>
     <ClCompile Include="Results.cpp">
+      <Filter>Source Files</Filter>
+    </ClCompile>
+    <ClCompile Include="MeasureResults.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
   </ItemGroup>
